Fix handling of OpenAPI nullable defined with `allOf` in renderer.
OpenAPI 3.0 markes an object as nullable by setting `nullable: true`.
This can be set directly on the object, e.g. `{ type: "string",
nullable: true }`, or it can be merged via `allOf`:

```json
{
  "allOf": [
    { "type": "string" },
    { "nullable": true },
  ]
}
```

The current fumadocs-openapi renderer treats `anyOf` as similar to a
TypeScript `&` operator. For example, the `anyOf` above would be
rendered as `string & null` when it should be `string | null`.