UI: fix navbar height for border
The height of the header overlaps overlaps main content by 1 pixel in mobile view and 2 pixels in desktop (border all around the header) view due to the fact that the border is not taken into account when setting the height of the nav in this component. I suggest this fix with hard coded heights but you are free to fix another way if you want.