fix: ensure metadata points to production url
This change ensures that links point to the right url on production env. e.g. if you look at the sitemap.xml file you can see that all urls in it are incorrectly pointed to the temporary vercel deployment url: https://fumadocs.dev/sitemap.xml
This will make the sitemap be pointed to the right url fumadocs.dev and in turn help google discover the website content.